+++ cockpit/langgraph/client-tools/python/src/graph.py
@@ -16,7 +16,7 @@ from langgraph.graph import StateGraph, END
 from langgraph.graph.message import add_messages
 from typing_extensions import Annotated, TypedDict
 
-from threadplane_client_tools import bind_client_tools
+from threadplane.client_tools import bind_client_tools
 
 PROMPTS_DIR = Path(__file__).parent.parent / "prompts"
