+++ packages/threadplane-client-tools/README.md
@@ -24,7 +24,7 @@ pip install threadplane-client-tools
 
 ```python
 from langgraph.graph import END, StateGraph
-from threadplane_client_tools import bind_client_tools, route_after_agent
+from threadplane.client_tools import bind_client_tools, route_after_agent
 
 # Server-side tools your graph owns
 SERVER_TOOLS = [search_tool, calculator_tool]
@@ -59,7 +59,7 @@ graph.add_conditional_edges("agent", router, {"tools": "tools", "__end__": END})
 ### Lower-level helpers
 
 ```python
-from threadplane_client_tools import (
+from threadplane.client_tools import (
     client_tool_specs,   # → list of OpenAI function-tool dicts
     client_tool_names,   # → set[str] of client tool names
     has_client_tool_call,  # → bool
