+++ packages/threadplane-client-tools/pyproject.toml
@@ -18,4 +18,4 @@ dependencies = [
 test = ["pytest>=8"]
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/threadplane_client_tools"]
+packages = ["src/threadplane"]
