+++ packages/threadplane/README.md
@@ -1,0 +1,14 @@
+# threadplane
+
+This distribution only reserves the `threadplane` top-level name on PyPI.
+
+It contains no importable code. The functional packages live under the
+`threadplane.*` namespace:
+
+- `threadplane-client-tools` — importable as `threadplane.client_tools`
+
+Install a functional package directly:
+
+```bash
+pip install threadplane-client-tools
+```
