+++ packages/threadplane/pyproject.toml
@@ -1,0 +1,14 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "threadplane"
+version = "0.0.0"
+description = "Namespace placeholder for the threadplane.* packages."
+readme = "README.md"
+license = { text = "MIT" }
+requires-python = ">=3.10"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/threadplane"]
